refactor: moved useLocalStorage state to TokenProvider component
