--- src/hooks/TokenContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const TokenContext = createContext();
-
-export default TokenContext;
